Moved string functions to Utility/String
