+++ src/Platform/Posix/Posix.cpp
@@ -14,7 +14,7 @@
 #include <filesystem>
 
 #include "Engine/Point.h"
-#include "Engine/Strings.h"
+#include "Utility/String.h"
 
 void OS_MsgBox(const char *msg, const char *title) {
     SDL_ShowSimpleMessageBox(SDL_MESSAGEBOX_ERROR, title, msg, nullptr);
