+++ src/Utility/String.h
